Moved bot events to Listeners
