+++ Bot_NetCore/Bot.cs
@@ -1,11 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Net;
 using System.Reflection;
 using System.Threading.Tasks;
-using System.Timers;
 using System.Xml;
 using System.Xml.Linq;
 using System.Xml.Serialization;
@@ -14,14 +11,10 @@ using Bot_NetCore.Entities;
 using Bot_NetCore.Misc;
 using DSharpPlus;
 using DSharpPlus.CommandsNext;
-using DSharpPlus.CommandsNext.Attributes;
-using DSharpPlus.CommandsNext.Exceptions;
 using DSharpPlus.Entities;
-using DSharpPlus.EventArgs;
 using DSharpPlus.Exceptions;
 using DSharpPlus.Interactivity;
 using DSharpPlus.Interactivity.Enums;
-using Microsoft.VisualBasic.FileIO;
 
 // ReSharper disable MemberCanBePrivate.Global
 // ReSharper disable UnassignedField.Global
@@ -37,15 +30,6 @@ namespace Bot_NetCore
         ///     Словарь, содержащий в качестве ключа пользователя Discord, а в качестве значения - время истечения кулдауна.
         /// </summary>
         public static Dictionary<DiscordUser, DateTime> ShipCooldowns = new Dictionary<DiscordUser, DateTime>();
-
-        /// <summary>
-        ///     Словарь, содержащий в качестве ключа пользователя Discord, а в качестве значения - время истечения кулдауна.
-        /// </summary>
-        public static Dictionary<DiscordUser, DateTime> EmojiCooldowns = new Dictionary<DiscordUser, DateTime>();
-
-        ///     Invites список приглашений
-        /// </summary>
-        public List<DiscordInvite> Invites;
 
         /// <summary>
         ///     DiscordClient бота.
@@ -102,7 +86,7 @@ namespace Bot_NetCore
             var cfg = new DiscordConfiguration
             {
                 Token = BotSettings.Token,
-                LogLevel = LogLevel.Info,
+                LogLevel = LogLevel.Debug,
                 AutoReconnect = true,
                 TokenType = TokenType.Bot,
                 UseInternalLogHandler = true
@@ -129,59 +113,19 @@ namespace Bot_NetCore
 
             Interactivity = Client.UseInteractivity(icfg);
 
+            //Команды
             Commands.RegisterCommands(Assembly.GetExecutingAssembly());
+
+            //Кастомнуа справка команд
             Commands.SetHelpFormatter<HelpFormatter>();
 
             //Ивенты
-            Client.Ready += ClientOnReady;
-
             AsyncListenerHandler.InstallListeners(Client, this);
 
-
-            //Не используются
-            //Client.MessageReactionRemoved += ClientOnMessageReactionRemoved; //Не нужный ивент
-
             //Логгер
-            Client.DebugLogger.LogMessageReceived += DebugLoggerOnLogMessageReceived;
-#if DEBUG
-            Client.ClientErrored += args =>
-            {
-                Console.WriteLine(args.Exception.InnerException);
-                return Task.CompletedTask;
-            };
-#endif
-
-            Commands.CommandErrored += CommandsOnCommandErrored;
-            Commands.CommandExecuted += CommandsOnCommandExecuted;
+            Client.DebugLogger.LogMessageReceived += Listeners.LoggerListener.LogOnLogMessageReceived;
 
             await Client.ConnectAsync();
-
-            //Таймер, который каждую минуту проверяет все баны и удаляет истёкшие.
-            var checkExpiredReports = new Timer(60000);
-            checkExpiredReports.Elapsed += CheckExpiredReports;
-            checkExpiredReports.AutoReset = true;
-            checkExpiredReports.Enabled = true;
-
-            //Таймер который каждую минуту проверяет истекшие сообщения в каналах
-            var clearChannelMessages = new Timer(60000);
-            clearChannelMessages.Elapsed += ClearChannelMessagesOnElapsed;
-            clearChannelMessages.AutoReset = true;
-            clearChannelMessages.Enabled = true;
-
-            var clearVotes = new Timer(60000);
-            clearVotes.Elapsed += ClearAndRepairVotesOnElapsed;
-            clearVotes.AutoReset = true;
-            clearVotes.Enabled = true;
-
-            var deleteShips = new Timer(60000 * 10);
-            deleteShips.Elapsed += DeleteShipsOnElapsed;
-            deleteShips.AutoReset = true;
-            deleteShips.Enabled = true;
-            
-            var clearSubscriptions = new Timer(60000);
-            clearSubscriptions.Elapsed += ClearSubscriptionsOnElapsed;
-            clearSubscriptions.AutoReset = true;
-            clearSubscriptions.Enabled = true;
 
             if (!Directory.Exists("generated")) Directory.CreateDirectory("generated");
             if (!File.Exists("generated/attachments_messages.csv")) File.Create("generated/attachments_messages.csv");
@@ -190,1237 +134,13 @@ namespace Bot_NetCore
             await Task.Delay(-1);
         }
 
-        private async void ClearSubscriptionsOnElapsed(object sender, ElapsedEventArgs e)
-        {
-            for (int i = 0; i < Subscriber.Subscribers.Count; ++i)
-            {
-                var sub = Subscriber.Subscribers.Values.ToArray()[i];
-                if (DateTime.Now > sub.SubscriptionEnd)
-                {
-                    try
-                    {
-                        var guild = Client.Guilds[BotSettings.Guild];
-                        var member = await guild.GetMemberAsync(sub.Member);
-                        if (member != null)
-                        {
-                            await member.SendMessageAsync("Ваша подписка истекла :cry:");
-                        }
-
-                        await DonatorCommands.DeletePrivateRoleAsync(guild, member.Id);
-
-                        Subscriber.Subscribers.Remove(sub.Member);
-                        Subscriber.Save(BotSettings.SubscriberXML);
-                    }
-                    catch (Exception ex)
-                    {
-                        Client.DebugLogger.LogMessage(LogLevel.Error, "Bot",
-                            $"Возникла ошибка при очистке подписок {ex.StackTrace}.",
-                            DateTime.Now);
-                    }
-                }
-            }
-        }
-
-        private async Task ClientOnGuildAvailable(GuildCreateEventArgs e)
-        {
-            await UpdateMembersCountAsync(Client, e.Guild.MemberCount);
-        }
-
-        private async void DeleteShipsOnElapsed(object sender, ElapsedEventArgs e)
-        {
-            for (int i = 0; i < ShipList.Ships.Count; ++i)
-            {
-                var ship = ShipList.Ships.Values.ToArray()[i];
-                if ((DateTime.Now - ship.LastUsed).Days >= 3)
-                {
-                    var channel = Client.Guilds[BotSettings.Guild].GetChannel(ship.Channel);
-
-                    ulong ownerId = 0;
-                    foreach (var member in ship.Members.Values)
-                        if (member.Type == MemberType.Owner)
-                        {
-                            ownerId = member.Id;
-                            break;
-                        }
-
-                    DiscordMember owner = null;
-                    try
-                    {
-                        owner = await Client.Guilds[BotSettings.Guild].GetMemberAsync(ownerId);
-                        await owner.SendMessageAsync(
-                            "Ваш приватный корабль был неактивен долгое время и поэтому он был удалён. \n**Пожалуйста, не отправляйте новый запрос на создание, если" +
-                            " не планируете пользоваться этой функцией**");
-                    }
-                    catch (NotFoundException)
-                    {
-                        // ничего не делаем, владелец покинул сервер
-                    }
-
-                    ship.Delete();
-                    ShipList.SaveToXML(Bot.BotSettings.ShipXML);
-
-                    await channel.DeleteAsync();
-
-                    var doc = XDocument.Load("actions.xml");
-                    foreach (var action in doc.Element("actions").Elements("action"))
-                        if (Convert.ToUInt64(action.Value) == ownerId)
-                            action.Remove();
-                    doc.Save("actions.xml");
-
-                    await Client.Guilds[BotSettings.Guild].GetChannel(BotSettings.ModlogChannel).SendMessageAsync(
-                        "**Удаление корабля**\n\n" +
-                        $"**Модератор:** {Client.CurrentUser}\n" +
-                        $"**Корабль:** {ship.Name}\n" +
-                        $"**Владелец:** {owner}\n" +
-                        $"**Дата:** {DateTime.Now}");
-                }
-            }
-        }
-
-        private async void ClearAndRepairVotesOnElapsed(object sender, ElapsedEventArgs e)
-        {
-            for (int i = 0; i < Vote.Votes.Count; ++i)
-            {
-                var vote = Vote.Votes.Values.ToArray()[i];
-                try
-                {
-                    var message = await Client.Guilds[BotSettings.Guild].GetChannel(BotSettings.VotesChannel)
-                        .GetMessageAsync(vote.Message);
-                    if (DateTime.Now >= vote.End && (DateTime.Now - vote.End).Days < 10) // выключение голосования
-                    {
-                        if (message.Reactions.Count == 0) continue;
-
-                        var author = await Client.Guilds[BotSettings.Guild].GetMemberAsync(vote.Author);
-                        var embed = Utility.GenerateVoteEmbed(
-                            author,
-                            vote.Yes > vote.No ? DiscordColor.Green : DiscordColor.Red,
-                            vote.Topic, vote.End,
-                            vote.Voters.Count,
-                            vote.Yes,
-                            vote.No,
-                            vote.Id);
-
-                        await message.ModifyAsync(embed: embed);
-                        await message.DeleteAllReactionsAsync();
-                    }
-                    else if (DateTime.Now >= vote.End && (DateTime.Now - vote.End).Days >= 3 && !message.Pinned) // архивирование голосования
-                    {
-                        var author = await Client.Guilds[BotSettings.Guild].GetMemberAsync(vote.Author);
-                        var embed = Utility.GenerateVoteEmbed(
-                            author,
-                            vote.Yes > vote.No ? DiscordColor.Green : DiscordColor.Red,
-                            vote.Topic, vote.End,
-                            vote.Voters.Count,
-                            vote.Yes,
-                            vote.No,
-                            vote.Id);
-
-                        var doc = new XDocument();
-                        var root = new XElement("Voters");
-                        foreach (var voter in vote.Voters)
-                            root.Add(new XElement("Voter", voter));
-                        doc.Add(root);
-                        doc.Save($"generated/voters-{vote.Id}.xml");
-
-                        var channel = Client.Guilds[BotSettings.Guild].GetChannel(BotSettings.VotesArchive);
-                        await channel.SendFileAsync($"generated/voters-{vote.Id}.xml", embed: embed);
-
-                        await message.DeleteAsync();
-                    }
-                    else if (DateTime.Now < vote.End) // починка голосования
-                    {
-                        if (message.Reactions.Count < 2)
-                        {
-                            await message.DeleteAllReactionsAsync();
-                            await message.CreateReactionAsync(DiscordEmoji.FromName(Client, ":white_check_mark:"));
-                            await message.CreateReactionAsync(DiscordEmoji.FromName(Client, ":no_entry:"));
-                        }
-                    }
-                }
-                catch (NotFoundException)
-                {
-                    //Do nothing, message not found
-                }
-                catch (Exception ex)
-                {
-                    Client.DebugLogger.LogMessage(LogLevel.Error, "Bot",
-                        $"Возникла ошибка при очистке голосований {ex.StackTrace}.",
-                        DateTime.Now);
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Очистка сообщений из каналов
-        /// </summary>
-        /// <param name="sender"></param>
-        /// <param name="e"></param>
-        private async void ClearChannelMessagesOnElapsed(object sender, ElapsedEventArgs e)
-        {
-            try
-            {
-                var guild = Client.Guilds[BotSettings.Guild];
-
-                var channels = new Dictionary<DiscordChannel, TimeSpan>
-                {
-                    { guild.GetChannel(BotSettings.FindChannel), new TimeSpan(0, 30, 0) },           //30 минут для канала поиска
-                    { guild.GetChannel(BotSettings.FleetCreationChannel), new TimeSpan(24, 0, 0) }   //24 часа для канала создания рейда
-                };
-
-                foreach (var channel in channels)
-                {
-                    try
-                    {
-                        var messages = await channel.Key.GetMessagesAsync();
-                        var toDelete = messages.ToList()
-                            .Where(x => !x.Pinned).ToList()                                                                           //Не закрепленные сообщения
-                            .Where(x => DateTimeOffset.Now.Subtract(x.CreationTimestamp.Add(channel.Value)).TotalSeconds > 0);     //Опубликованные ранее определенного времени
-
-                        if (toDelete.Count() > 0)
-                        {
-                            await channel.Key.DeleteMessagesAsync(toDelete);
-                            Client.DebugLogger.LogMessage(LogLevel.Info, "Bot", $"Канал {channel.Key.Name} был очищен.", DateTime.Now);
-                        }
-
-                    }
-                    catch (Exception ex)
-                    {
-                        Client.DebugLogger.LogMessage(LogLevel.Warning, "Bot", $"Ошибка при удалении сообщений в {channel.Key.Name}. \n{ex.Message}", DateTime.Now);
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                Client.DebugLogger.LogMessage(LogLevel.Error, "Bot", $"Ошибка при удалении сообщений в каналах. \n" +
-                        $"{ex.GetType()}" +
-                        $"{ex.Message}\n" +
-                        $"{ex.StackTrace}",
-                    DateTime.Now);
-            }
-        }
-
-        private Task CommandsOnCommandExecuted(CommandExecutionEventArgs e)
-        {
-            var command = (e.Command.Parent != null ? e.Command.Parent.Name + " " : "") + e.Command.Name;
-
-            e.Context.Client.DebugLogger.LogMessage(LogLevel.Info,
-                    "Bot",
-                    $"Пользователь {e.Context.Member.Username}#{e.Context.Member.Discriminator} ({e.Context.Member.Id}) выполнил команду {command}",
-                    DateTime.Now);
-            return Task.CompletedTask; //Пришлось добавить, выдавало ошибку при компиляции
-        }
-
-#nullable enable //Выдавало warning
-        private async void DebugLoggerOnLogMessageReceived(object? sender, DebugLogMessageEventArgs e)
-        {
-            if (!Directory.Exists("logs")) Directory.CreateDirectory("logs");
-
-            var fileName = "logs/" + DateTime.Today.ToString("yyyy-MM-dd");
-
-            var loglevel = "";
-            switch (e.Level)
-            {
-                case LogLevel.Critical:
-                    loglevel = "Critical";
-                    break;
-                case LogLevel.Error:
-                    loglevel = "Error";
-                    break;
-                case LogLevel.Warning:
-                    loglevel = "Warning";
-                    break;
-                case LogLevel.Info:
-                    loglevel = "Info";
-                    break;
-                case LogLevel.Debug:
-                    loglevel = "Debug";
-                    break;
-            }
-
-            //файл для удобного парсинга
-            using (var fs = new FileStream(fileName + ".csv", FileMode.Append))
-            {
-                using (var sw = new StreamWriter(fs))
-                {
-                    var message = e.Message.Replace("\"", "'");
-                    await sw.WriteLineAsync($"{e.Timestamp:s},{loglevel},{e.Application},\"{message}\"");
-                }
-            }
-
-            //файл для удобного просмотра
-            using (var fs = new FileStream(fileName + ".log", FileMode.Append))
-            {
-                using (var sw = new StreamWriter(fs))
-                {
-                    await sw.WriteLineAsync($"[{e.Timestamp:G}] [{loglevel}] [{e.Application}] {e.Message}");
-                }
-            }
-        }
-#nullable disable
-
-        private async void CheckExpiredReports(object sender, ElapsedEventArgs e)
-        {
-            var guild = await Client.GetGuildAsync(BotSettings.Guild);
-
-            //Check for expired bans
-            var toUnban = from ban in BanList.BannedMembers.Values
-                          where ban.UnbanDateTime <= DateTime.Now
-                          select ban;
-
-            if (toUnban.Count() > 0)
-            {
-                foreach (var ban in toUnban)
-                {
-                    try
-                    {
-                        await guild.UnbanMemberAsync(ban.Id);
-                    }
-                    catch (NotFoundException)
-                    {
-                        //пользователь мог и не быть заблокирован через Discord
-                    }
-
-                    ban.Unban();
-
-                    var user = await Client.GetUserAsync(ban.Id);
-                    await guild.GetChannel(BotSettings.ModlogChannel).SendMessageAsync(
-                        "**Снятие Бана**\n\n" +
-                        $"**Модератор:** {Client.CurrentUser.Username}\n" +
-                        $"**Пользователь:** {user}\n" +
-                        $"**Дата:** {DateTime.Now}\n");
-
-                    Client.DebugLogger.LogMessage(LogLevel.Info, "Bot", $"Пользователь {user} был разбанен.", DateTime.Now);
-                }
-
-                BanList.SaveToXML(BotSettings.BanXML);
-
-                Client.DebugLogger.LogMessage(LogLevel.Info, "Bot", "Бан-лист был обновлён.", DateTime.Now);
-            }
-
-            //Check for expired mutes
-            var count = ReportList.Mutes.Count;
-            ReportList.Mutes.Values.Where(x => x.Expired()).ToList()
-                .ForEach(async x =>
-                {
-                    ReportList.Mutes.Remove(x.Id);
-                    try
-                    {
-                        var user = await guild.GetMemberAsync(x.Id);
-                        await user.RevokeRoleAsync(guild.GetRole(BotSettings.MuteRole), "Unmuted");
-                    }
-                    catch (NotFoundException)
-                    {
-                        //Пользователь не найден
-                    }
-                });
-            if (count != ReportList.Mutes.Count)
-                ReportList.SaveToXML(BotSettings.ReportsXML);
-
-            //Check for expired voice mutes
-            count = ReportList.VoiceMutes.Count;
-            ReportList.VoiceMutes.Values.Where(x => x.Expired()).ToList()
-                .ForEach(async x =>
-                {
-                    ReportList.VoiceMutes.Remove(x.Id);
-                    try
-                    {
-                        var user = await guild.GetMemberAsync(x.Id);
-                        await user.RevokeRoleAsync(guild.GetRole(BotSettings.VoiceMuteRole), "Unmuted");
-                    }
-                    catch (NotFoundException)
-                    {
-                        //Пользователь не найден
-                    }
-                });
-            if (count != ReportList.VoiceMutes.Count)
-                ReportList.SaveToXML(BotSettings.ReportsXML);
-        }
-
-        private async Task ClientOnMessageReactionRemoved(MessageReactionRemoveEventArgs e)
-        {
-            if (e.User.IsBot) return;
-
-            //Проверка если сообщение с принятием правил
-            if (e.Message.Id == BotSettings.CodexMessageId)
-            {
-                //При надобности добавить кулдаун
-                //if (EmojiCooldowns.ContainsKey(e.User)) // проверка на кулдаун
-                //    if ((EmojiCooldowns[e.User] - DateTime.Now).Seconds > 0) return;
-
-                //// если проверка успешно пройдена, добавим пользователя
-                //// в словарь кулдаунов
-                //EmojiCooldowns[e.User] = DateTime.Now.AddSeconds(BotSettings.FastCooldown);
-
-                //Забираем роль
-                var user = (DiscordMember)e.User;
-                if (user.Roles.Any(x => x.Id == BotSettings.CodexRole))
-                    await user.RevokeRoleAsync(e.Channel.Guild.GetRole(BotSettings.CodexRole));
-
-                return;
-            }
-
-            //Emissary Message
-            if (e.Message.Id == BotSettings.EmissaryMessageId) return;
-        }
-
-        private async Task ClientOnMessageReactionAdded(MessageReactionAddEventArgs e)
-        {
-            if (e.User.IsBot) return;
-
-            //Проверка если сообщение с принятием правил
-            if (e.Message.Id == BotSettings.CodexMessageId && e.Emoji.GetDiscordName() == ":white_check_mark:")
-            {
-                //При надобности добавить кулдаун
-                /*if (EmojiCooldowns.ContainsKey(e.User)) // проверка на кулдаун
-                    if ((EmojiCooldowns[e.User] - DateTime.Now).Seconds > 0) return;
-
-                // если проверка успешно пройдена, добавим пользователя
-                // в словарь кулдаунов
-                EmojiCooldowns[e.User] = DateTime.Now.AddSeconds(BotSettings.FastCooldown);*/
-
-                //Проверка на purge
-                if (ReportList.CodexPurges.ContainsKey(e.User.Id))
-                    if (!ReportList.CodexPurges[e.User.Id].Expired()) //Проверка истекшей блокировки
-                    {
-                        var moderator = await e.Channel.Guild.GetMemberAsync(ReportList.CodexPurges[e.User.Id].Moderator);
-                        try
-                        {
-                            await ((DiscordMember)e.User).SendMessageAsync(
-                                "**Возможность принять правила заблокирована**\n" +
-                                $"**Снятие через:** {Utility.FormatTimespan(ReportList.CodexPurges[e.User.Id].getRemainingTime())}\n" +
-                                $"**Модератор:** {moderator.Username}#{moderator.Discriminator}\n" +
-                                $"**Причина:** {ReportList.CodexPurges[e.User.Id].Reason}\n");
-                        }
-
-                        catch (UnauthorizedException)
-                        {
-                            //user can block the bot
-                        }
-                        return;
-                    }
-                    else
-                        ReportList.CodexPurges.Remove(e.User.Id); //Удаляем блокировку если истекла
-
-                //Выдаем роль правил
-                var user = (DiscordMember)e.User;
-                if (!user.Roles.Contains(e.Channel.Guild.GetRole(BotSettings.CodexRole)))
-                {
-                    //Выдаем роль правил
-                    await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.CodexRole));
-
-                    //Убираем роль блокировки правил
-                    await user.RevokeRoleAsync(e.Channel.Guild.GetRole(BotSettings.PurgeCodexRole));
-
-                    e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                        $"Пользователь {e.User.Username}#{e.User.Discriminator} ({e.User.Id}) подтвердил прочтение правил через реакцию.",
-                        DateTime.Now);
-                }
-
-                return;
-            }
-
-            //Проверка если сообщение с принятием правил
-            if (e.Message.Id == BotSettings.FleetCodexMessageId && e.Emoji.GetDiscordName() == ":white_check_mark:")
-            {
-                //Проверка на purge
-                if (ReportList.FleetPurges.ContainsKey(e.User.Id))
-                    if (!ReportList.FleetPurges[e.User.Id].Expired()) //Проверка истекшей блокировки
-                    {
-                        var moderator = await e.Channel.Guild.GetMemberAsync(ReportList.FleetPurges[e.User.Id].Moderator);
-                        try
-                        {
-                            await ((DiscordMember)e.User).SendMessageAsync(
-                                "**Возможность принять правила рейда заблокирована**\n" +
-                                $"**Снятие через:** {Utility.FormatTimespan(ReportList.FleetPurges[e.User.Id].getRemainingTime())}\n" +
-                                $"**Модератор:** {moderator.Username}#{moderator.Discriminator}\n" +
-                                $"**Причина:** {ReportList.FleetPurges[e.User.Id].Reason}\n");
-                        }
-
-                        catch (UnauthorizedException)
-                        {
-                            //user can block the bot
-                        }
-                        return;
-                    }
-                    else
-                        ReportList.FleetPurges.Remove(e.User.Id); //Удаляем блокировку если истекла
-
-                //Выдаем роль правил рейда
-                var user = (DiscordMember)e.User;
-                if (!user.Roles.Any(x => x.Id == BotSettings.FleetCodexRole))
-                {
-                    await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.FleetCodexRole));
-                    e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                        $"Пользователь {e.User.Username}#{e.User.Discriminator} ({e.User.Id}) подтвердил прочтение правил рейда.",
-                        DateTime.Now);
-                }
-
-                return;
-            }
-
-            //Проверка на сообщение эмиссарства
-            if (e.Message.Id == BotSettings.EmissaryMessageId)
-            {
-                await e.Message.DeleteReactionAsync(e.Emoji, e.User);
-
-                if (EmojiCooldowns.ContainsKey(e.User)) // проверка на кулдаун
-                    if ((EmojiCooldowns[e.User] - DateTime.Now).Seconds > 0) return;
-
-                // если проверка успешно пройдена, добавим пользователя
-                // в словарь кулдаунов
-                EmojiCooldowns[e.User] = DateTime.Now.AddSeconds(BotSettings.FastCooldown);
-
-                //Проверка у пользователя уже существующих ролей эмисарства и их удаление
-                var user = (DiscordMember)e.User;
-                user.Roles.Where(x => x.Id == BotSettings.EmissaryGoldhoadersRole ||
-                                x.Id == BotSettings.EmissaryTradingCompanyRole ||
-                                x.Id == BotSettings.EmissaryOrderOfSoulsRole ||
-                                x.Id == BotSettings.EmissaryAthenaRole ||
-                                x.Id == BotSettings.EmissaryReaperBonesRole ||
-                                x.Id == BotSettings.HuntersRole ||
-                                x.Id == BotSettings.ArenaRole).ToList()
-                         .ForEach(async x => await user.RevokeRoleAsync(x));
-
-                //Выдаем роль в зависимости от реакции
-                switch (e.Emoji.GetDiscordName())
-                {
-                    case ":moneybag:":
-                        await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.EmissaryGoldhoadersRole));
-                        break;
-                    case ":pig:":
-                        await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.EmissaryTradingCompanyRole));
-                        break;
-                    case ":skull:":
-                        await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.EmissaryOrderOfSoulsRole));
-                        break;
-                    case ":gem:":
-                        await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.EmissaryAthenaRole));
-                        break;
-                    case ":bone:":
-                        await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.EmissaryReaperBonesRole));
-                        break;
-                    case ":fish:":
-                        await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.HuntersRole));
-                        break;
-                    case ":axe:":
-                        await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.ArenaRole));
-                        break;
-                    default:
-                        break;
-                }
-                //Отправка в лог
-                e.Client.DebugLogger.LogMessage(LogLevel.Info, "SoT",
-                    $"{e.User.Username}#{e.User.Discriminator} получил новую роль эмиссарства.",
-                    DateTime.Now);
-
-                return;
-            }
-
-            //Проверка на голосование
-            if (e.Message.Channel.Id == BotSettings.VotesChannel)
-            {
-                var vote = Vote.GetByMessageId(e.Message.Id);
-
-                await e.Message.DeleteReactionAsync(e.Emoji, e.User);
-
-                // Проверка на окончание голосования
-                if (DateTime.Now > vote.End)
-                {
-                    return;
-                }
-
-                // Проверка на предыдущий голос
-                if (vote.Voters.ContainsKey(e.User.Id))
-                {
-                    return;
-                }
-
-                if (e.Emoji.GetDiscordName() == ":white_check_mark:")
-                {
-                    vote.Voters.Add(e.User.Id, true);
-                    ++vote.Yes;
-                }
-                else
-                {
-                    vote.Voters.Add(e.User.Id, false);
-                    ++vote.No;
-                }
-
-                var total = vote.Voters.Count;
-
-                var author = await e.Guild.GetMemberAsync(vote.Author);
-                var embed = Utility.GenerateVoteEmbed(
-                    author,
-                    DiscordColor.Yellow,
-                    vote.Topic,
-                    vote.End,
-                    vote.Voters.Count,
-                    vote.Yes,
-                    vote.No,
-                    vote.Id);
-
-                Vote.Save(BotSettings.VotesXML);
-
-                await e.Message.ModifyAsync(embed: embed);
-                await (await e.Guild.GetMemberAsync(e.User.Id)).SendMessageAsync($"{BotSettings.OkEmoji} Спасибо, ваш голос учтён!");
-            }
-
-            //then check if it is a private ship confirmation message
-            foreach (var ship in ShipList.Ships.Values)
-            {
-                if (ship.Status) continue;
-
-                if (e.Message.Id == ship.CreationMessage)
-                {
-                    if (e.Emoji == DiscordEmoji.FromName((DiscordClient)e.Client, ":white_check_mark:"))
-                    {
-                        var name = ship.Name;
-                        var channel = await e.Channel.Guild.CreateChannelAsync($"☠{name}☠", ChannelType.Voice,
-                            e.Channel.Guild.GetChannel(BotSettings.PrivateCategory), bitrate: BotSettings.Bitrate);
-
-                        var member = await e.Channel.Guild.GetMemberAsync(ShipList.Ships[name].Members.ToArray()[0].Value.Id);
-
-                        await channel.AddOverwriteAsync(member, Permissions.UseVoice);
-                        await channel.AddOverwriteAsync(e.Channel.Guild.GetRole(BotSettings.CodexRole), Permissions.AccessChannels);
-                        await channel.AddOverwriteAsync(e.Channel.Guild.EveryoneRole, Permissions.None, Permissions.UseVoice);
-
-                        ShipList.Ships[name].SetChannel(channel.Id);
-                        ShipList.Ships[name].SetStatus(true);
-                        ShipList.Ships[name].SetMemberStatus(member.Id, true);
-
-                        ShipList.SaveToXML(BotSettings.ShipXML);
-
-                        await member.SendMessageAsync(
-                            $"{BotSettings.OkEmoji} Запрос на создание корабля **{name}** был подтвержден администратором **{e.User.Username}#{e.User.Discriminator}**");
-                        await e.Channel.SendMessageAsync(
-                            $"{BotSettings.OkEmoji} Администратор **{e.User.Username}#{e.User.Discriminator}** успешно подтвердил " +
-                            $"запрос на создание корабля **{name}**!");
-
-                        e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                            $"Администратор {e.User.Username}#{e.User.Discriminator} ({e.User.Id}) подтвердил создание приватного корабля {name}.",
-                            DateTime.Now);
-                    }
-                    else if (e.Emoji == DiscordEmoji.FromName((DiscordClient)e.Client, ":no_entry:"))
-                    {
-                        var name = ship.Name;
-                        var member =
-                            await e.Channel.Guild.GetMemberAsync(ShipList.Ships[name].Members.ToArray()[0].Value.Id);
-
-                        ShipList.Ships[name].Delete();
-                        ShipList.SaveToXML(BotSettings.ShipXML);
-
-                        var doc = XDocument.Load("actions.xml");
-                        foreach (var action in doc.Element("actions").Elements("action"))
-                            if (Convert.ToUInt64(action.Value) == member.Id)
-                                action.Remove();
-                        doc.Save("actions.xml");
-
-                        await member.SendMessageAsync(
-                            $"{BotSettings.OkEmoji} Запрос на создание корабля **{name}** был отклонен администратором **{e.User.Username}#{e.User.Discriminator}**");
-                        await e.Channel.SendMessageAsync(
-                            $"{BotSettings.OkEmoji} Администратор **{e.User.Username}#{e.User.Discriminator}** успешно отклонил запрос на " +
-                            $"создание корабля **{name}**!");
-
-                        e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                            $"Администратор {e.User.Username}#{e.User.Discriminator} ({e.User.Id}) отклонил создание приватного корабля {name}.",
-                            DateTime.Now);
-                    }
-                    else
-                    {
-                        return;
-                    }
-                }
-            }
-
-        }
-
-        /// <summary>
-        ///     Новый обработчик команд.
-        /// </summary>
-        private async Task ClientOnMessageCreated(MessageCreateEventArgs e)
-        {
-            if (e.Channel.Id == BotSettings.CodexReserveChannel)
-            {
-                if (!IsModerator(await e.Guild.GetMemberAsync(e.Author.Id)))
-                    await e.Message.DeleteAsync();
-
-                //Проверка на purge
-                if (ReportList.CodexPurges.ContainsKey(e.Author.Id))
-                    if (!ReportList.CodexPurges[e.Author.Id].Expired()) //Проверка истекшей блокировки
-                    {
-                        var moderator = await e.Channel.Guild.GetMemberAsync(ReportList.CodexPurges[e.Author.Id].Moderator);
-                        try
-                        {
-                            await ((DiscordMember)e.Author).SendMessageAsync(
-                                "**Возможность принять правила заблокирована**\n" +
-                                $"**Снятие через:** {Utility.FormatTimespan(ReportList.CodexPurges[e.Author.Id].getRemainingTime())}\n" +
-                                $"**Модератор:** {moderator.Username}#{moderator.Discriminator}\n" +
-                                $"**Причина:** {ReportList.CodexPurges[e.Author.Id].Reason}\n");
-                        }
-
-                        catch (UnauthorizedException)
-                        {
-                            //user can block the bot
-                        }
-                        return;
-                    }
-                    else
-                        ReportList.CodexPurges.Remove(e.Author.Id); //Удаляем блокировку если истекла
-
-                //Выдаем роль правил
-                var user = (DiscordMember)e.Author;
-                if (!user.Roles.Contains(e.Channel.Guild.GetRole(BotSettings.CodexRole)))
-                {
-                    //Выдаем роль правил
-                    await user.GrantRoleAsync(e.Channel.Guild.GetRole(BotSettings.CodexRole));
-
-                    //Убираем роль блокировки правил
-                    await user.RevokeRoleAsync(e.Channel.Guild.GetRole(BotSettings.PurgeCodexRole));
-
-                    e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                        $"Пользователь {e.Author.Username}#{e.Author.Discriminator} ({e.Author.Id}) подтвердил прочтение правил через сообщение.",
-                        DateTime.Now);
-                }
-            }
-
-            if (e.Message.Attachments.Count > 0 && !e.Message.Author.IsBot)
-            {
-                var message = $"**Автор:** {e.Message.Author}\n" +
-                              $"**Канал:**  {e.Message.Channel}\n" +
-                              $"**Сообщение:** {e.Message.Id}\n" +
-                              $"**Вложение:**\n";
-
-                using (var client = new WebClient())
-                {
-                    var attachment = e.Message.Attachments[0]; //проверить: не может быть больше 1 вложения в сообщении
-                    var file = $"generated/attachments/{attachment.FileName}";
-                    client.DownloadFile(attachment.Url, file);
-                    var logMessage = await e.Guild.GetChannel(BotSettings.AttachmentsLog).SendFileAsync(file, message);
-                    File.Delete(file);
-
-                    using (var fs = new FileStream("generated/attachments_messages.csv", FileMode.Append))
-                    using (var sw = new StreamWriter(fs))
-                        await sw.WriteLineAsync($"{e.Message.Id},{logMessage.Id}");
-                }
-            }
-
-            if (e.Message.Content.StartsWith("> "))
-                if (IsModerator(await e.Guild.GetMemberAsync(e.Author.Id)))
-                {
-                    var messageStrings = e.Message.Content.Split('\n');
-                    var command = "";
-                    foreach (var str in messageStrings)
-                        if (str.StartsWith("<@"))
-                        {
-                            command = str;
-                            break;
-                        }
-
-                    var args = command.Split(' ');
-                    var receiver = args[0];
-                    var action = args[1];
-
-                    switch (action)
-                    {
-                        case "w":
-                            await e.Message.DeleteAsync();
-                            RunCommand((DiscordClient)e.Client, CommandType.Warn, args, e.Message);
-                            return;
-                        default:
-                            return;
-                    }
-                }
-        }
-
-        /// <summary>
-        ///     Лог посещений
-        /// </summary>
-        private async Task ClientOnGuildMemberRemoved(GuildMemberRemoveEventArgs e)
-        {
-            // Сохранение ролей участника
-            var roles = e.Member.Roles;
-            var rolesToSave = new List<ulong>();
-            var ignoredRoles = new List<ulong> //роли, которые не нужно сохранять при выходе
-            {
-                BotSettings.CodexRole,
-                BotSettings.FleetCodexRole,
-                BotSettings.MuteRole,
-                BotSettings.VoiceMuteRole,
-                BotSettings.EmissaryAthenaRole,
-                BotSettings.EmissaryGoldhoadersRole,
-                BotSettings.EmissaryReaperBonesRole,
-                BotSettings.EmissaryTradingCompanyRole,
-                BotSettings.EmissaryOrderOfSoulsRole,
-                e.Guild.EveryoneRole.Id,
-            };
-
-            foreach (var role in roles)
-            {
-                if (!ignoredRoles.Contains(role.Id))
-                {
-                    rolesToSave.Add(role.Id);
-                }
-            }
-
-            if (rolesToSave.Count != 0)
-            {
-                UsersLeftList.Users[e.Member.Id] = new UserLeft(e.Member.Id, rolesToSave);
-                UsersLeftList.SaveToXML(BotSettings.UsersLeftXML);
-            }
-
-            await e.Guild.GetChannel(BotSettings.UserlogChannel)
-                .SendMessageAsync(
-                    $"**Участник покинул сервер:** {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}).");
-
-            //Обновляем статус бота
-            await UpdateMembersCountAsync(e.Client, e.Guild.MemberCount);
-
-            //Если пользователь не был никем приглашен, то при выходе он будет сохранен.
-            if (!InviterList.Inviters.ToList().Any(i => i.Value.Referrals.ContainsKey(e.Member.Id)))
-            {
-                if (!InviterList.Inviters.ContainsKey(0))
-                    InviterList.Inviters[0] = new Inviter(0, false);
-
-                InviterList.Inviters[0].AddReferral(e.Member.Id, false);
-            }
-
-            //При выходе обновляем реферала на неактив.
-            InviterList.Inviters.ToList().Where(i => i.Value.Referrals.ContainsKey(e.Member.Id)).ToList()
-                                         .ForEach(i => i.Value.UpdateReferral(e.Member.Id, false));
-
-            InviterList.SaveToXML(BotSettings.InviterXML);
-
-            await LeaderboardCommands.UpdateLeaderboard(e.Guild);
-
-            e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                $"Участник {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}) покинул сервер.",
-                DateTime.Now);
-        }
-
-        /// <summary>
-        ///     Приветственное сообщение + лог посещений + проверка на бан
-        /// </summary>
-        private async Task ClientOnGuildMemberAdded(GuildMemberAddEventArgs e)
-        {
-            //Проверка на бан
-            if (BanList.BannedMembers.ContainsKey(e.Member.Id) && BanList.BannedMembers[e.Member.Id].UnbanDateTime > BanList.BannedMembers[e.Member.Id].BanDateTime)
-            {
-                var ban = BanList.BannedMembers[e.Member.Id];
-                //Отправка сообщения в лс
-                try
-                {
-                    await e.Member.SendMessageAsync($"Вы были заблокированы на этом сервере. **Причина:** " +
-                                                $"{ban.Reason}. **Блокировка истекает:** ${ban.UnbanDateTime}.");
-                }
-                catch (UnauthorizedException)
-                {
-                    //user can block the bot
-                }
-                await e.Member.BanAsync(0, "Autoban");
-
-                return;
-            }
-
-            //Проверка на mute
-            if (ReportList.Mutes.ContainsKey(e.Member.Id) && !ReportList.Mutes[e.Member.Id].Expired())
-            {
-                //Отправка сообщения в лс
-                try
-                {
-                    await e.Member.SendMessageAsync(
-                        $"**Вам выдан мут на данном сервере**\n\n" +
-                        $"**Снятие через:** {Utility.FormatTimespan(ReportList.Mutes[e.Member.Id].getRemainingTime())}\n" +
-                        $"**Причина:** {ReportList.Mutes[e.Member.Id].Reason}");
-                }
-                catch (UnauthorizedException)
-                {
-                    //user can block the bot
-                }
-
-                //Выдаем роль мута
-                await e.Member.GrantRoleAsync(e.Guild.GetRole(BotSettings.MuteRole));
-            }
-
-            //Проверка на voice mute
-            if (ReportList.VoiceMutes.ContainsKey(e.Member.Id) && !ReportList.VoiceMutes[e.Member.Id].Expired())
-            {
-                //Отправка сообщения в лс
-                try
-                {
-                    await e.Member.SendMessageAsync(
-                        $"**Вам выдан голосовой мут на данном сервере**\n\n" +
-                        $"**Снятие через:** {Utility.FormatTimespan(ReportList.VoiceMutes[e.Member.Id].getRemainingTime())}\n" +
-                        $"**Причина:** {ReportList.VoiceMutes[e.Member.Id].Reason}");
-                }
-                catch (UnauthorizedException)
-                {
-                    //user can block the bot
-                }
-
-                //Выдаем роль мута
-                await e.Member.GrantRoleAsync(e.Guild.GetRole(BotSettings.VoiceMuteRole));
-            }
-
-            //Проверка на purge
-            if (ReportList.CodexPurges.ContainsKey(e.Member.Id) && !ReportList.CodexPurges[e.Member.Id].Expired())
-                await e.Member.GrantRoleAsync(e.Guild.GetRole(BotSettings.PurgeCodexRole));
-
-            //Выдача доступа к приватным кораблям
-            try
-            {
-                var ships = ShipList.Ships.Values.Where(x => x.Members.ContainsKey(e.Member.Id));
-                foreach (var ship in ships)
-                    await e.Guild.GetChannel(ship.Channel).AddOverwriteAsync(e.Member, Permissions.UseVoice);
-            }
-            catch (Exception ex)
-            {
-                e.Client.DebugLogger.LogMessage(LogLevel.Warning, "Bot",
-                   $"Ошибка при выдаче доступа к приватному кораблю. \n{ex.Message}\n{ex.StackTrace}",
-                   DateTime.Now); ;
-            }
-
-            var invites = Invites.AsReadOnly().ToList(); //Сохраняем список старых инвайтов в локальную переменную
-            var guildInvites = await e.Guild.GetInvitesAsync(); //Запрашиваем новый список инвайтов
-            Invites = guildInvites.ToList(); //Обновляю список инвайтов
-
-            try
-            {
-                await e.Member.SendMessageAsync($"**Привет, {e.Member.Mention}!\n**" +
-                                                "Мы рады что ты присоединился к нашему сообществу :wink:!\n\n" +
-                                                "Прежде чем приступать к игре, прочитай и прими правила в канале " +
-                                                "`#👮-пиратский-кодекс-👮`. После принятия можешь ознакомиться с гайдом по боту" +
-                                                "в канале `#📚-гайд-по-боту-📚`.\n" +
-                                                "Если у тебя есть какие-то вопросы, не стесняйся писать администрации.\n\n" +
-                                                "**Удачной игры!**");
-            }
-            catch (UnauthorizedException)
-            {
-                //Пользователь заблокировал бота
-            }
-
-            // Выдача ролей, которые были у участника перед выходом.
-            if (UsersLeftList.Users.ContainsKey(e.Member.Id))
-            {
-                foreach (var role in UsersLeftList.Users[e.Member.Id].Roles)
-                {
-                    try
-                    {
-                        await e.Member.GrantRoleAsync(e.Guild.GetRole(role));
-                    }
-                    catch (NotFoundException)
-                    {
-
-                    }
-                }
-
-                UsersLeftList.Users.Remove(e.Member.Id);
-                UsersLeftList.SaveToXML(BotSettings.UsersLeftXML);
-            }
-
-            try
-            {
-                //Находит обновившийся инвайт по количеству приглашений
-                //Вызывает NullReferenceException в случае если ссылка только для одного использования
-                var updatedInvite = guildInvites.ToList().Find(g => invites.Find(i => i.Code == g.Code).Uses < g.Uses);
-
-                //Если не удалось определить инвайт, значит его нет в новых так как к.во использований ограничено и он был удален
-                if (updatedInvite == null)
-                {
-                    updatedInvite = invites.Where(p => guildInvites.All(p2 => p2.Code != p.Code))                       //Ищем удаленный инвайт
-                                           .Where(x => (x.CreatedAt.AddSeconds(x.MaxAge) < DateTimeOffset.Now))      //Проверяем если он не истёк
-                                           .FirstOrDefault();                                                           //С такими условиями будет только один такой инвайт
-                }
-
-                if (updatedInvite != null)
-                {
-
-                    await e.Guild.GetChannel(BotSettings.UserlogChannel)
-                    .SendMessageAsync(
-                        $"**Участник присоединился:** {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}) используя " +
-                        $"приглашение {updatedInvite.Code} от участника {updatedInvite.Inviter.Username}#{updatedInvite.Inviter.Discriminator}. ");
-
-                    e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                        $"Участник {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}) присоединился к серверу. Приглашение: {updatedInvite.Code} от участника {updatedInvite.Inviter.Username}#{updatedInvite.Inviter.Discriminator}.",
-                        DateTime.Now);
-
-                    //Проверяем если пригласивший уже существует, если нет то создаем
-                    if (!InviterList.Inviters.ContainsKey(updatedInvite.Inviter.Id))
-                        Inviter.Create(updatedInvite.Inviter.Id, e.Member.IsBot);
-
-                    //Проверяем если пользователь был ранее приглашен другими и обновляем активность, если нет то вносим в список
-                    if (InviterList.Inviters.ToList().Exists(x => x.Value.Referrals.ContainsKey(e.Member.Id)))
-                        InviterList.Inviters.ToList().Where(x => x.Value.Referrals.ContainsKey(e.Member.Id)).ToList()
-                            .ForEach(x => x.Value.UpdateReferral(e.Member.Id, true));
-                    else
-                        InviterList.Inviters[updatedInvite.Inviter.Id].AddReferral(e.Member.Id);
-
-                    InviterList.SaveToXML(BotSettings.InviterXML);
-                    //Обновление статистики приглашений
-                    await LeaderboardCommands.UpdateLeaderboard(e.Guild);
-                }
-                else
-                {
-                    await e.Guild.GetChannel(BotSettings.UserlogChannel)
-                        .SendMessageAsync(
-                            $"**Участник присоединился:** {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}). " +
-                            $"Приглашение не удалось определить.");
-
-                    e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                        $"Участник {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}) присоединился к серверу. Приглашение не удалось определить.",
-                        DateTime.Now);
-                }
-            }
-            catch (Exception ex)
-            {
-                await e.Guild.GetChannel(BotSettings.UserlogChannel)
-                    .SendMessageAsync(
-                        $"**Участник присоединился:** {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}). " +
-                        $"При попытке отследить инвайт произошла ошибка.");
-
-                e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                    $"Участник {e.Member.Username}#{e.Member.Discriminator} ({e.Member.Id}) присоединился к серверу. Приглашение не удалось определить.",
-                    DateTime.Now);
-
-                e.Client.DebugLogger.LogMessage(LogLevel.Warning, "Bot",
-                    "Не удалось определить приглашение.",
-                    DateTime.Now);
-
-                var errChannel = e.Guild.GetChannel(BotSettings.ErrorLog);
-
-                var message = "**Ошибка при логгинге инвайта**\n" +
-                              $"**Пользователь:** {e.Member}\n" +
-                              $"**Исключение:** {ex.GetType()}:{ex.Message}\n" +
-                              $"**Трассировка стека:** \n```{ex.StackTrace}```\n" +
-                              $"{ex}";
-
-                await errChannel.SendMessageAsync(message);
-            }
-
-            //Обновляем статус бота
-            await UpdateMembersCountAsync(e.Client, e.Guild.MemberCount);
-        }
-
-        /// <summary>
-        ///     Система автосоздания кораблей
-        /// </summary>
-        private async Task ClientOnVoiceStateUpdated(VoiceStateUpdateEventArgs e)
-        {
-            try
-            {
-                if (e.Channel.Id == BotSettings.AutocreateGalleon ||
-                    e.Channel.Id == BotSettings.AutocreateBrigantine ||
-                    e.Channel.Id == BotSettings.AutocreateSloop
-                ) // мы создаем канал, если пользователь зашел в один из каналов автосоздания
-                {
-                    if (ShipCooldowns.ContainsKey(e.User)) // проверка на кулдаун
-                        if ((ShipCooldowns[e.User] - DateTime.Now).Seconds > 0)
-                        {
-                            var m = await e.Guild.GetMemberAsync(e.User.Id);
-                            await m.PlaceInAsync(e.Guild.GetChannel(BotSettings.WaitingRoom));
-                            await m.SendMessageAsync($"{BotSettings.ErrorEmoji} Вам нужно подождать " +
-                                                     $"**{(ShipCooldowns[e.User] - DateTime.Now).Seconds}** секунд прежде чем " +
-                                                     "создавать новый корабль!");
-                            e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                                $"Участник {e.User.Username}#{e.User.Discriminator} ({e.User.Discriminator}) был перемещён в комнату ожидания.",
-                                DateTime.Now);
-                            return;
-                        }
-
-                    // если проверка успешно пройдена, добавим пользователя
-                    // в словарь кулдаунов
-                    ShipCooldowns[e.User] = DateTime.Now.AddSeconds(BotSettings.FastCooldown);
-
-                    //Проверка на эмиссарство
-                    var channelSymbol = BotSettings.AutocreateSymbol;
-                    ((DiscordMember)e.User).Roles.ToList().ForEach(x =>
-                    {
-                        if (x.Id == BotSettings.EmissaryGoldhoadersRole)
-                            channelSymbol = DiscordEmoji.FromName((DiscordClient)e.Client, ":moneybag:");
-                        else if (x.Id == BotSettings.EmissaryTradingCompanyRole)
-                            channelSymbol = DiscordEmoji.FromName((DiscordClient)e.Client, ":pig:");
-                        else if (x.Id == BotSettings.EmissaryOrderOfSoulsRole)
-                            channelSymbol = DiscordEmoji.FromName((DiscordClient)e.Client, ":skull:");
-                        else if (x.Id == BotSettings.EmissaryAthenaRole)
-                            channelSymbol = DiscordEmoji.FromName((DiscordClient)e.Client, ":gem:");
-                        else if (x.Id == BotSettings.EmissaryReaperBonesRole)
-                            channelSymbol = DiscordEmoji.FromName((DiscordClient)e.Client, ":skull_crossbones:");
-                        else if (x.Id == BotSettings.HuntersRole)
-                            channelSymbol = DiscordEmoji.FromName((DiscordClient)e.Client, ":fish:");
-                        else if (x.Id == BotSettings.ArenaRole)
-                            channelSymbol = DiscordEmoji.FromName((DiscordClient)e.Client, ":crossed_swords:");
-
-                    });
-
-                    DiscordChannel created = null;
-                    // Проверяем канал в котором находится пользователь
-
-                    if (e.Channel.Id == BotSettings.AutocreateSloop) //Шлюп
-                        created = await e.Guild.CreateChannelAsync(
-                            $"{channelSymbol} Шлюп {e.User.Username}", ChannelType.Voice,
-                            e.Guild.GetChannel(BotSettings.AutocreateCategory), bitrate: BotSettings.Bitrate, userLimit: 2);
-                    else if (e.Channel.Id == BotSettings.AutocreateBrigantine) // Бригантина
-                        created = await e.Guild.CreateChannelAsync(
-                            $"{channelSymbol} Бриг {e.User.Username}", ChannelType.Voice,
-                            e.Guild.GetChannel(BotSettings.AutocreateCategory), bitrate: BotSettings.Bitrate, userLimit: 3);
-                    else // Галеон
-                        created = await e.Guild.CreateChannelAsync(
-                            $"{channelSymbol} Галеон {e.User.Username}", ChannelType.Voice,
-                            e.Guild.GetChannel(BotSettings.AutocreateCategory), bitrate: BotSettings.Bitrate, userLimit: 4);
-
-                    var member = await e.Guild.GetMemberAsync(e.User.Id);
-
-                    await member.PlaceInAsync(created);
-
-                    e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot",
-                        $"Участник {e.User.Username}#{e.User.Discriminator} ({e.User.Id}) создал канал через автосоздание.",
-                        DateTime.Now);
-                }
-                else if (e.Channel.Id == BotSettings.FindShip)
-                {
-                    var shipCategory = e.Guild.GetChannel(BotSettings.AutocreateCategory);
-
-                    var membersLookingForTeam = new List<ulong>();
-                    foreach (var message in (await e.Guild.GetChannel(BotSettings.FindChannel).GetMessagesAsync(100)))
-                    {
-                        if (message.Pinned) continue; // автор закрепленного сообщения не должен учитываться
-                        if (membersLookingForTeam.Contains(message.Author.Id)) continue; // автор сообщения уже мог быть добавлен в лист
-
-                        membersLookingForTeam.Add(message.Author.Id);
-                    }
-
-                    var possibleChannels = new List<DiscordChannel>();
-                    foreach (var ship in shipCategory.Children)
-                        if (ship.Users.Count() < ship.UserLimit)
-                            foreach (var user in ship.Users)
-                                if (membersLookingForTeam.Contains(user.Id))
-                                    possibleChannels.Add(ship);
-
-                    var m = await e.Guild.GetMemberAsync(e.User.Id);
-                    if (possibleChannels.Count == 0)
-                    {
-                        await m.PlaceInAsync(e.Guild.GetChannel(BotSettings.WaitingRoom));
-                        await m.SendMessageAsync($"{BotSettings.ErrorEmoji} Не удалось найти подходящий корабль.");
-                        return;
-                    }
-
-                    var random = new Random();
-                    var rShip = random.Next(0, possibleChannels.Count);
-
-                    await m.PlaceInAsync(possibleChannels[rShip]);
-                    e.Client.DebugLogger.LogMessage(LogLevel.Info, "Bot", $"Пользователь {m.Username}#{m.Discriminator} успешно воспользовался поиском корабля!", DateTime.Now);
-                    return;
-                }
-                else if (e.Channel.ParentId == BotSettings.PrivateCategory)
-                {
-                    foreach (var ship in ShipList.Ships.Values)
-                    {
-                        if (ship.Channel == e.Channel.Id)
-                        {
-                            ship.SetLastUsed(DateTime.Now);
-                            ShipList.SaveToXML(BotSettings.ShipXML);
-                            break;
-                        }
-                    }
-                }
-            }
-            catch (NullReferenceException) // исключение выбрасывается если пользователь покинул канал
-            {
-                // нам здесь ничего не надо делать, просто пропускаем
-            }
-
-            try
-            {
-                // удалим пустые каналы
-
-                var autocreatedChannels = new List<DiscordChannel>(); // это все автосозданные каналы
-                foreach (var channel in e.Guild.Channels.Values)
-                    if (channel.Type == ChannelType.Voice
-                        && channel.ParentId == BotSettings.AutocreateCategory)
-                        autocreatedChannels.Add(channel);
-
-                var notEmptyChannels = new List<DiscordChannel>(); // это все НЕ пустые каналы
-                foreach (var voiceState in e.Guild.VoiceStates.Values) notEmptyChannels.Add(voiceState.Channel);
-
-                var forDeletionChannels = autocreatedChannels.Except(notEmptyChannels); // это пустые каналы
-                foreach (var channel in forDeletionChannels) await channel.DeleteAsync(); // мы их удаляем
-            }
-            catch (NotFoundException) // Если пользователь пересоздает канал перейдя с уже автосозданного канала
-            {
-                // пропускаем
-            }
-
-            //Проверка на пустые рейды
-            if (e.Before != null && e.Before.Channel != null)
-            {
-                var leftChannel = e.Before.Channel;
-
-                //Пользователь вышел из автоматически созданных каналов рейда
-                if (leftChannel.Parent.Name.StartsWith("Рейд") &&
-                   leftChannel.ParentId != BotSettings.FleetCategory &&
-                   !leftChannel.Users.Contains(e.User))
-                {
-                    //Проверка всех каналов рейда на присутствие в них игроков
-                    var fleetIsEmpty = leftChannel.Parent.Children
-                                            .Where(x => x.Type == ChannelType.Voice)
-                                            .Where(x => x.Users.Count() > 0)
-                                            .Count() == 0;
-
-                    //Удаляем каналы и категорию
-                    if (fleetIsEmpty)
-                    {
-                        foreach (var emptyChannel in leftChannel.Parent.Children)
-                            await emptyChannel.DeleteAsync();
-                        await leftChannel.Parent.DeleteAsync();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        ///     Проверка на создание инвайтов
-        /// </summary>
-        private async Task ClientOnInviteCreated(InviteCreateEventArgs e)
-        {
-            var guildInvites = await e.Client.Guilds[BotSettings.Guild].GetInvitesAsync();
-            Invites = guildInvites.ToList();
-        }
-
-        /// <summary>
-        ///     Проверка на удаление инвайтов
-        /// </summary>
-        private async Task ClientOnInviteDeleted(InviteDeleteEventArgs e)
-        {
-            var guildInvites = await e.Client.Guilds[BotSettings.Guild].GetInvitesAsync();
-            Invites = guildInvites.ToList();
-        }
-
-        /// <summary>
-        ///     Сообщение в лог о готовности клиента
-        /// </summary>
-        private async Task ClientOnReady(ReadyEventArgs e)
-        {
-            e.Client.DebugLogger.LogMessage(LogLevel.Info, "SoT", $"Sea Of Thieves Bot, version {BotSettings.Version}",
-                DateTime.Now);
-            e.Client.DebugLogger.LogMessage(LogLevel.Info, "SoT", "Made by Actis",
-                DateTime.Now); // и еще немного ЧСВ
-
-            var guild = e.Client.Guilds[BotSettings.Guild];
-
-            var member = await guild.GetMemberAsync(e.Client.CurrentUser.Id);
-            await member.ModifyAsync(x => x.Nickname = $"SeaOfThieves {BotSettings.Version}");
-
-            var guildInvites = await guild.GetInvitesAsync();
-            Invites = guildInvites.ToList();
-        }
-
-
         /// <summary>
         ///     Обновляет статус бота, отображая к.во участников на сервере.
         /// </summary>
         /// <param name="client">Клиент бота</param>
         /// <param name="number">Число пользователей на сервере</param>
         /// <returns></returns>
-        public async Task UpdateMembersCountAsync(DiscordClient client, int number)
+        public static async Task UpdateMembersCountAsync(DiscordClient client, int number)
         {
             await client.UpdateStatusAsync(new DiscordActivity($"за {number} пользователями", ActivityType.Watching));
         }
@@ -1494,7 +214,7 @@ namespace Bot_NetCore
             return false;
         }
 
-        private static async void RunCommand(DiscordClient client, CommandType type, string[] args,
+        public static async void RunCommand(DiscordClient client, CommandType type, string[] args,
             DiscordMessage message)
         {
             switch (type)
