+++ Bot_NetCore/Listeners/Template.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net;
+using System.Reflection;
+using System.Threading.Tasks;
+using System.Timers;
+using System.Xml;
+using System.Xml.Linq;
+using System.Xml.Serialization;
+using Bot_NetCore.Commands;
+using Bot_NetCore.Entities;
+using Bot_NetCore.Misc;
+using DSharpPlus;
+using DSharpPlus.CommandsNext;
+using DSharpPlus.CommandsNext.Attributes;
+using DSharpPlus.CommandsNext.Exceptions;
+using DSharpPlus.Entities;
+using DSharpPlus.EventArgs;
+using DSharpPlus.Exceptions;
+using DSharpPlus.Interactivity;
+using DSharpPlus.Interactivity.Enums;
+using Microsoft.VisualBasic.FileIO;
+
+namespace Bot_NetCore.Listeners
+{
+    public static class TemplateListener
+    {
+        [AsyncListener(EventTypes.Ready)]
+        public static async Task OnEvent(ReadyEventArgs e)
+        {
+            await Task.CompletedTask;
+        }
+    }
+}
